refactor(eager): merge BS mixin into engine file (one BS module)
The mixin split (b280340) put the eager-BS propagator methods in a new
_torch_eager_boundary.py, leaving two confusingly-named BS files (the
_eager_boundary_saving.py engine + the _torch_eager_boundary.py mixin).
Fold _EagerBoundarySavingMixin into _eager_boundary_saving.py — its
natural home: the methods construct ReconState, dispatch _REVERSE_DRIVERS
and call _BoundarySaveStep, all defined there, so the same-module lazy
imports collapse to direct refs.  Delete the twin file; _PropTorchEager
imports the mixin from the engine module.  (custom_grad keeps its own
mixin file — it has no separate engine module.)

Verified: import/MRO unchanged, eager BS pytest 27/27, 4-way eB/eF
cosine 1.0 across all equations.
